+++ engine/core/src/definitions/tests.rs
@@ -4,6 +4,7 @@
 
 mod caps;
 mod mapping;
+mod size;
 
 use super::*;